requirements.txt
# Data manipulation and analysis
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Data visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Record Linkage
recordlinkage>=0.15.0
dedupe>=2.0.0

# Machine Learning
scikit-learn>=1.3.0
torch>=2.0.0
transformers>=4.30.0

# Data quality
jellyfish>=0.11.0  # String similarity
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.23.0
ipywidgets>=8.0.0

# Progress bars
tqdm>=4.65.0

# Database (optional)
sqlalchemy>=2.0.0

# Kaggle API (for downloading datasets)
kaggle>=1.5.0

# Data validation
great-expectations>=0.17.0

# Label Studio (optional, for manual annotation)
label-studio>=1.8.0
